move package to source
