+++ app/src/main/java/com/github/satoshun/reactivex/exoplayer2/sample/MainActivity.kt
@@ -6,7 +6,7 @@ import android.support.v7.app.AppCompatActivity
 import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
-import com.github.satoshun.reactivex.exoplayer2.hls.events
+import com.github.satoshun.reactivex.exoplayer2.source.events
 import com.google.android.exoplayer2.DefaultLoadControl
 import com.google.android.exoplayer2.DefaultRenderersFactory
 import com.google.android.exoplayer2.ExoPlayerFactory
