+++ exoplayer/src/main/java/com/github/satoshun/reactivex/exoplayer2/source/RxMediaSourceEvent.kt
@@ -1,4 +1,4 @@
-package com.github.satoshun.reactivex.exoplayer2.hls
+package com.github.satoshun.reactivex.exoplayer2.source
 
 import android.os.Handler
 import android.support.annotation.CheckResult
